requirements.txt
dnspython
Flask
Flask-Cors
Flask-Mail
flask-ngrok
Flask-Session
Flask-SocketIO
gunicorn
numpy
openpyxl
pandas
phonenumbers
pycountry
PyJWT
pymongo
python-jose
requests
eventlet
gevent-websocket
python-dotenv
pillow